Scraper no longer collects a clue's value, only the question and answer.
The value is not an intrinsic part of a clue and is only important if
the goal of the scraper is to save clues within the context of a
Jeopardy game that aired on tv. Because the purpose of the scraper is to
save clues to create custom trivia games, where a
value can be dynamically assigned to a clue, saving the value of a clue
as it appears on JArchive is unecessary.